fix(e2e): split e2eBundledExploreRejectionDiagnostics out of shared library (Refs #2336)
`app/integration_test/e2e_test_shared.dart` had ticked over the 1000
non-comment-line cap enforced by `repo.dart_file_non_comment_line_size`
(`SPEC/program/dart-file-non-comment-line-size.md`,
`colonizethis-code-review.mdc`) at 1004 lines after lifting the
`e2eTap{New,Old}WorldRegionTab` helpers in 82e3af2ea, blocking the
`Quality (lint, analyze, observer)` job on PR #2722.

Mirror the existing `e2e_test_shared_panels.dart` /
`e2e_test_shared_bootstrap.dart` split: extract the largest cohesive
block — the cold-failure `e2eBundledExploreRejectionDiagnostics`
helper plus its `provinceHasFoggedOrBetter` closure — into a new
sub-library `e2e_test_shared_diagnostics.dart`. The parent file now
also re-exports it (`export 'e2e_test_shared_diagnostics.dart';`) so
the AC1 barrel `e2e_helpers.dart` keeps resolving the public name
without import-site churn. The `OrderEngine` / `MoveOrder` /
`WorkOrder` / `Unit` / `kUnitTypeExplorer` / `suggestWorkOrders`
imports moved with the function (no other helper in the parent file
referenced them).

Verification:
- `dart run tool/check_dart_file_non_comment_line_size.dart` ->
  `no violations found.` (parent file now well under 1000).
- `cd app && flutter analyze integration_test/...` -> `No issues found!`
- `cd app && flutter test test/e2e_bundled_explore_rejection_diagnostics_test.dart \\
   test/e2e_helpers_barrel_test.dart test/e2e_test_shared_smoke_test.dart \\
   test/e2e_tap_region_tab_test.dart test/e2e_region_chip_selected_test.dart`
  -> 77/77 pass (diagnostics pin + barrel tear-offs unchanged).
- `cd app && flutter test test/e2e_explore_assign_enabled_from_civilian_snapshot_test.dart \\
   test/e2e_non_home_human_fleet_in_new_world_from_ct_snapshot_test.dart \\
   test/e2e_non_home_human_fleet_in_coastal_new_world_sea_from_ct_snapshot_test.dart \\
   test/e2e_player_has_any_new_world_fogged_or_better_from_ct_snapshot_test.dart \\
   test/e2e_naval_panel_shows_non_home_fleet_in_new_world_test.dart \\
   test/e2e_radio_list_tiles_in_alert_dialogs_test.dart \\
   test/e2e_dismiss_transient_ui_test.dart` -> 95/95 pass.

No SPEC change — pure file-size split, every public symbol keeps its
identifier, signature, and call-site contract.

Refs #2336

Made with [Cursor](https://cursor.com)
